Fix : 챌린지 신청하기 모듈화 추가수정
